conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
adwaita-icon-theme=40.1.1=hca03da5_1
atk-1.0=2.36.0=h7fe96df_0
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cairo=1.16.0=h29d4eff_2
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
dbus=1.13.18=h8280c03_0
defusedxml=0.7.1=pyhd8ed1ab_0
docopt=0.6.2=pypi_0
epoxy=1.5.4=h1a28f6b_2
expat=2.4.4=hc377ac9_0
ffmpeg=4.3.2=ha8417bf_1
fontconfig=2.13.1=h1f4a5ec_0
fpdf2=2.5.6=pyhd8ed1ab_0
freetype=2.11.0=h1192e45_0
fribidi=1.0.10=h1a28f6b_0
gdk-pixbuf=2.42.6=h74ef11a_4
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
glib=2.70.2=hccf11d3_4
glib-tools=2.70.2=hccf11d3_4
gmp=6.2.1=h9f76cd9_0
gnutls=3.6.13=h706517b_1
gobject-introspection=1.68.0=py39h3459c9f_3
graphite2=1.3.14=hc377ac9_1
gtk3=3.24.21=h321fa23_1
harfbuzz=2.8.1=hffeda63_0
hdf5=1.10.6=nompi_h0fc092c_1114
hicolor-icon-theme=0.17=hca03da5_2
icu=68.1=hc377ac9_0
idna=3.4=pypi_0
imutils=0.5.4=pypi_0
jasper=2.0.33=hc3cd1e9_1
jpeg=9e=h1a28f6b_0
krb5=1.19.2=h3b8d789_0
lame=3.100=h1a8c8d9_1003
lcms2=2.12=had6a04f_0
libblas=3.9.0=13_osxarm64_openblas
libcblas=3.9.0=13_osxarm64_openblas
libcurl=7.81.0=h8fe1914_0
libcxx=12.0.0=hf6beb65_1
libedit=3.1.20210910=h1a28f6b_0
libev=4.33=h642e427_1
libevent=2.1.12=hf27765b_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=12_2_0_hd922786_31
libgfortran5=12.2.0=h0eea778_31
libglib=2.70.2=h67e64d8_4
libiconv=1.16=h1a28f6b_2
liblapack=3.9.0=13_osxarm64_openblas
liblapacke=3.9.0=13_osxarm64_openblas
libnghttp2=1.46.0=he4cd7f6_0
libopenblas=0.3.18=openmp_h5dd58f0_0
libopencv=4.5.1=py39h9e916e0_1
libpng=1.6.37=hb8d0fd4_0
libpq=12.9=h65cfe13_3
libprotobuf=3.15.8=hccf11d3_1
librsvg=2.50.7=h90a7944_0
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=hb80f160_3
libtiff=4.2.0=h01837e1_1
libwebp=1.2.2=h68602c7_0
libwebp-base=1.2.2=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=12.0.0=haf9daa7_1
lz4-c=1.9.3=hc377ac9_0
ncurses=6.3=h07bb92c_1
nettle=3.6=hc6a1b29_0
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numpy=1.21.3=py39h1f3b974_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.1=py39hdf13c20_1
openh264=2.1.1=habe5f53_0
openjpeg=2.4.0=h062765e_1
openssl=1.1.1t=h03a7124_0
packaging=23.0=pyhd8ed1ab_0
pango=1.48.7=h79d33a7_0
pcre=8.45=hc377ac9_0
pillow=8.2.0=py39h282ff46_1
pip=23.0.1=pyhd8ed1ab_0
pipreqs=0.4.11=pypi_0
pixman=0.40.0=h1a28f6b_0
py-opencv=4.5.1=py39hfa6204d_1
pymongo=3.12.0=py39hfb83b0d_0
pyqt=5.15.2=py39he8f2410_0
pyqt5-sip=4.19.25=pypi_0
pyqtchart=5.15.2=pypi_0
pytesseract=0.3.10=pyhd8ed1ab_0
python=3.9.12=hbdb9e5c_1
python_abi=3.9=2_cp39
qt=5.15.2=h4e944ae_2
readline=8.1.2=h46ed386_0
requests=2.28.2=pypi_0
setuptools=67.6.0=pyhd8ed1ab_0
sqlite=3.38.5=h1058600_0
svg.path=6.2=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tzdata=2022g=h191b570_0
urllib3=1.26.15=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
x264=1!161.3030=h3422bc3_1
xz=5.2.6=h57fd34a_0
yarg=0.1.9=pypi_0
zlib=1.2.13=h03a7124_4
zstd=1.5.2=h8574219_0
